Added .clang-tidy's and moved unity into separate folder
