+++ .clang-tidy
@@ -1,0 +1,1 @@
+HeaderFilterRegex: ''
